Cambiado Integer por Long para los id's

+++ src/main/java/com/fot/backcontask/service/store/StoreService.java
@@ -3,6 +3,6 @@ package com.fot.backcontask.service.store;
 import com.fot.backcontask.model.Store;
 import com.fot.backcontask.service.Service;
 
-public interface StoreService extends Service<Store, Integer> {
+public interface StoreService extends Service<Store, Long> {
 
 }
